chore(workspace): rename the app publishable applications data source
